Removing Copy-CSShadowCopyItem and Enable-CSRdp
A decision was made for CIMSweep to only pull information via WMI/CIM w/
Get- functions. CIMSweep must not modify or change any configurations on
targets if it is to be trusted as a clean auditing tool.

Thank you so much, @xorrior for these contributions and I'm sorry I had
to drop them!